+++ lib/core/extensions/delta_extension_utils.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: always_specify_types
+
 import 'dart:convert';
 
 import 'package:collection/collection.dart';